Renamed 'ruleFailureReasons' to 'eligibilityRuleFailureReasons' for improved clarity
